Move test files in the js/src/data folder to the 'test' sub-folder
